pyproject.toml
[project]
name = "mtahir9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mutagen>=1.47.0",
    "pygame-ce>=2.5.6",
]
